refactor: modularise routes.py by moving large functions into controllers.py - helps improve code organisation and maintainability
