+++ app/routes.py
@@ -11,6 +11,14 @@ from app.processing_layer import CarbonFootprintCalculator
 from flask_login import login_user, logout_user, current_user, login_required
 from app.forms import LoginForm, SignupForm, ChangePasswordForm, ShareForm, ResetPasswordRequestForm, ResetPasswordForm, DeleteAccountForm, EditNameForm, EditEMailForm
 from app.email_utils import send_confirmation_email, send_password_reset_email, send_email_update_confirmation
+from app.controllers import (
+    process_travel_data,
+    process_vehicles_data,
+    process_home_data,
+    process_food_data,
+    process_shopping_data,
+    delete_user_and_data
+)
 
 @main.route('/')
 @login_required
@@ -160,88 +168,6 @@ def add_data():
         is_new_user=is_new_user
     )
 
-def process_travel_data(mode, pts:PublicTransitSimpleForm, pta:PublicTransitAdvancedForm, ats:AirTravelSimpleForm, ata:AirTravelAdvancedForm):
-    travel = Travel()
-    if mode == 'simple':
-        travel.public_transit_distance = pts.distance.data
-        travel.air_travel_distance = ats.distance.data
-    else:
-        travel.bus_kms = pta.bus_kms.data
-        travel.transit_rail_kms = pta.transit_rail_kms.data
-        travel.commuter_rail_kms = pta.commuter_rail_kms.data
-        travel.intercity_rail_kms = pta.intercity_rail_kms.data
-        travel.short_flights = ata.short_flights.data
-        travel.medium_flights = ata.medium_flights.data
-        travel.long_flights = ata.long_flights.data
-        travel.extended_flights = ata.extended_flights.data
-    return travel
-
-def process_vehicles_data(travel_id):
-    """
-    Process vehicle form data and return a list of Vehicle objects.
-    """
-    vehicles = []
-    valid_fuel_types = ['gasoline', 'diesel', 'electric']
-    i = 0
-
-    while True:
-        fuel_type = request.form.get(f'vehicle-fuel_type{i}')
-        distance = request.form.get(f'vehicle-distance{i}')
-        efficiency = request.form.get(f'vehicle-fuel_efficiency{i}')
-        if not fuel_type or fuel_type not in valid_fuel_types:
-            break
-        vehicles.append(Vehicle(
-            travel_id=travel_id,
-            fuel_type=fuel_type,
-            distance=distance,
-            fuel_efficiency=efficiency
-        ))
-        i += 1
-    return vehicles
-
-def process_home_data(form:HomeEnergyForm):
-    return Home(
-        electricity=form.electricity.data,
-        electricity_unit=form.electricity_unit.data,
-        electricity_frequency=form.electricity_frequency.data,
-        clean_energy_percentage=form.clean_energy_percentage.data,
-        natural_gas=form.natural_gas.data,
-        natural_gas_unit=form.natural_gas_unit.data,
-        natural_gas_frequency=form.natural_gas_frequency.data,
-        heating_oil=form.heating_oil.data,
-        heating_oil_unit=form.heating_oil_unit.data,
-        heating_oil_frequency=form.heating_oil_frequency.data,
-        living_space=form.living_space.data,
-        water_usage=form.water_usage.data
-    )
-
-def process_food_data(form:FoodForm):
-    return Food(
-        meat_fish_eggs=form.meat_fish_eggs.data,
-        grains_baked_goods=form.grains_baked_goods.data,
-        dairy=form.dairy.data,
-        fruits_vegetables=form.fruits_vegetables.data,
-        snacks_drinks=form.snacks_drinks.data
-    )
-
-def process_shopping_data(mode, simple_form:ShoppingSimpleForm, advanced_form:ShoppingAdvancedForm):
-    shopping = Shopping()
-    if mode == 'simple':
-        shopping.goods_multiplier = simple_form.goods_multiplier.data
-        shopping.services_multiplier = simple_form.services_multiplier.data
-    else:
-        shopping.furniture_appliances = advanced_form.furniture_appliances.data
-        shopping.clothing = advanced_form.clothing.data
-        shopping.entertainment = advanced_form.entertainment.data
-        shopping.office_supplies = advanced_form.office_supplies.data
-        shopping.personal_care = advanced_form.personal_care.data
-        shopping.services_food = advanced_form.services_food.data
-        shopping.education = advanced_form.education.data
-        shopping.communication = advanced_form.communication.data
-        shopping.loan = advanced_form.loan.data
-        shopping.transport = advanced_form.transport.data
-    return shopping
-
 @main.route('/view_data')
 @login_required
 def view_data():
@@ -291,7 +217,6 @@ def view_data():
                            locked=locked,
                            shared_with_me=shared_with_me_sorted)
 
-
 @main.route('/api/compare_emissions')
 @login_required
 def compare_emissions():
@@ -319,8 +244,6 @@ def compare_emissions():
         'other_emissions': get_grouped_data(other_emission),
         'other_name': f"{other_user.first_name} {other_user.last_name}"
     })
-
-
 
 @main.route('/api/emissions', methods=['GET'])
 @login_required
@@ -529,7 +452,6 @@ def share():
                            shopping_pct=shopping_pct,
                            total_emission=total_emission,
                            is_new_user=is_new_user)
-
 
 @main.route('/api/emissions/<email>', methods=['GET'])
 @login_required
@@ -928,40 +850,6 @@ def change_password():
                 flash(f"{form[field].label.text}: {error}", "error")
         
         return redirect(url_for('main.profile'))
-
-#helper function to delete any related data to user 
-def delete_user_and_data(user):
-    # Delete emissions and footprint-related models
-    footprints = CarbonFootprint.query.filter_by(user_id=user.id).all()
-    for fp in footprints:
-        # Delete emissions first
-        Emissions.query.filter_by(carbon_footprint_id=fp.id).delete()
-
-        # Delete linked travel and vehicles
-        if fp.travel_id:
-            Vehicle.query.filter_by(travel_id=fp.travel_id).delete()
-            Travel.query.filter_by(id=fp.travel_id).delete()
-
-        # Delete other linked models
-        if fp.home_id:
-            Home.query.filter_by(id=fp.home_id).delete()
-        if fp.food_id:
-            Food.query.filter_by(id=fp.food_id).delete()
-        if fp.shopping_id:
-            Shopping.query.filter_by(id=fp.shopping_id).delete()
-
-    # Delete the footprints
-    CarbonFootprint.query.filter_by(user_id=user.id).delete()
-
-    # Delete share records (both sent and received)
-    Share.query.filter(
-        (Share.from_user_id == user.id) |
-        (Share.to_user_id == user.id)
-    ).delete()
-
-    # Finally, delete the user
-    db.session.delete(user)
-    db.session.commit()
 
 @main.route('/delete_account', methods=['POST'])
 @login_required
